+++ DigitaleBriefwahlTests/YesNoElectionModelTests.cs
@@ -12,13 +12,19 @@ namespace DigitaleBriefwahlTests
 	[TestFixture]
 	public class YesNoElectionModelTests
 	{
-		[TestCase("E\nN\nJ\nN", ExpectedResult = "Election\n" +
+		[TestCase("E\nN\nJ\nN", false, ExpectedResult = "Election\n" +
 "--------\n" +
 "1. [E] Mickey Mouse\n" +
 "2. [N] Donald Duck\n" +
 "3. [J] Dagobert Duck\n" +
 "4. [N] Daisy Duck\n")]
-		public string GetResult(string votes)
+		[TestCase("E\nE\nE\nE", true, ExpectedResult = "Election\n" +
+"--------\n" +
+"1. [E] Mickey Mouse\n" +
+"2. [E] Donald Duck\n" +
+"3. [E] Dagobert Duck\n" +
+"4. [E] Daisy Duck\n")]
+		public string GetResult_TwoVotes(string votes, bool writeEmptyBallot)
 		{
 			// Setup
 			const string ini = @"[Election]
@@ -34,7 +40,32 @@ Kandidat4=Daisy Duck
 			var model = ElectionModelFactory.Create("Election", data) as YesNoElectionModel;
 
 			// Execute
-			return model.GetResult(votes.Split('\n').ToList());
+			return model.GetResult(votes.Split('\n').ToList(), writeEmptyBallot);
+		}
+
+		[TestCase("J", ExpectedResult = "Election\n" +
+"--------\n" +
+"1. [J] Mickey Mouse\n")]
+		[TestCase("E", ExpectedResult = "Election\n" +
+"--------\n" +
+"1. [E] Mickey Mouse\n")]
+		[TestCase("N", ExpectedResult = "Election\n" +
+"--------\n" +
+"1. [N] Mickey Mouse\n")]
+		public string GetResult_OneVote(string votes)
+		{
+			// Setup
+			const string ini = @"[Election]
+Text=Some description
+Typ=YesNo
+Stimmen=1
+Kandidat1=Mickey Mouse
+";
+			var data = ElectionModelTests.ReadIniDataFromString(ini);
+			var model = ElectionModelFactory.Create("Election", data) as YesNoElectionModel;
+
+			// Execute
+			return model.GetResult(votes.Split('\n').ToList(), false);
 		}
 
 		[TestCase("", typeof(ArgumentException))]
@@ -59,7 +90,7 @@ Kandidat4=Daisy Duck
 			var model = ElectionModelFactory.Create("Election", data) as YesNoElectionModel;
 
 			// Execute
-			Assert.That(() => model.GetResult(votes.Split('\n').ToList()),
+			Assert.That(() => model.GetResult(votes.Split('\n').ToList(), false),
 				Throws.Exception.TypeOf(expectedException));
 		}
 	}